Audio Rework
The Audio / Sound method has changed, it is now a browser source element in the "obs/sounds" folder.

Changes include the files moving to "obs/sounds/files" and the handling of the audio queue is similar to the notifications widget.

Through this, it can play far more file formats and is more stream-lined.